pyproject.toml
[tool.poetry]
name = "python-urlpinger"
version = "0.1.0"
description = ""
authors = ["Tim Bryant <timothybryant3@gmail.com>"]
readme = "README.md"
packages = [{include = "urlpinger", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
structlog = "^24.4.0"
asyncio = "^3.4.3"
aiohttp = "^3.10.5"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
asyncpg = "^0.29.0"
greenlet = "^3.1.0"
alembic = "^1.13.2"
prometheus-client = "^0.21.0"
tenacity = "^9.0.0"
apprise = "^1.9.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.1"
black = "^24.8.0"
pytest = "^8.3.3"
pre-commit = "^3.8.0"
isort = "^5.13.2"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
  ".", "urlpinger"
]
filterwarnings = [
    "error",
    "ignore::RuntimeWarning",
    "ignore::DeprecationWarning",
]
